backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
Pillow==10.4.0
python-socketio==5.14.2
email-validator==2.1.0
argon2-cffi==23.1.0
psycopg2-binary